+++ ran-booster-bitbucket.php
@@ -30,6 +30,8 @@ add_action(
 			|| 5 !== RAN_BOOSTER_PROVIDER_API_VERSION
 			|| ! defined( 'RAN_BOOSTER_LOGGING_API_VERSION' )
 			|| 1 !== RAN_BOOSTER_LOGGING_API_VERSION
+			|| ! defined( 'RAN_BOOSTER_ADDON_API_VERSION' )
+			|| 7 !== RAN_BOOSTER_ADDON_API_VERSION
 			|| ! $registry instanceof \RAN\RepositoryProvider\ProviderRegistry ) {
 			return;
 		}
@@ -58,10 +60,31 @@ add_action(
 );
 
 add_action(
+	'ran_booster_documentation_after_provider_bb',
+	static function ( string $documentationUrl, string $scope ): void {
+		if ( ! defined( 'RAN_BOOSTER_PROVIDER_API_VERSION' )
+			|| 5 !== RAN_BOOSTER_PROVIDER_API_VERSION
+			|| ! defined( 'RAN_BOOSTER_LOGGING_API_VERSION' )
+			|| 1 !== RAN_BOOSTER_LOGGING_API_VERSION
+			|| ! defined( 'RAN_BOOSTER_ADDON_API_VERSION' )
+			|| 7 !== RAN_BOOSTER_ADDON_API_VERSION ) {
+			return;
+		}
+
+		unset( $documentationUrl, $scope );
+
+		require __DIR__ . '/views/documentation.php';
+	},
+	10,
+	2
+);
+
+add_action(
 	'admin_notices',
 	static function (): void {
 		if ( defined( 'RAN_BOOSTER_PROVIDER_API_VERSION' ) && 5 === RAN_BOOSTER_PROVIDER_API_VERSION
-			&& defined( 'RAN_BOOSTER_LOGGING_API_VERSION' ) && 1 === RAN_BOOSTER_LOGGING_API_VERSION ) {
+			&& defined( 'RAN_BOOSTER_LOGGING_API_VERSION' ) && 1 === RAN_BOOSTER_LOGGING_API_VERSION
+			&& defined( 'RAN_BOOSTER_ADDON_API_VERSION' ) && 7 === RAN_BOOSTER_ADDON_API_VERSION ) {
 			return;
 		}